refactor: ViewingRooms section
